Codificação de senhas (shs.h) corrigido para compilar em 64 bits. Concluído editor de texto do MUD. Melhor organização dos comandos de edição do MUD.

+++ shs.h
@@ -16,7 +16,7 @@
 #define SHS_H
 
 typedef unsigned char BYTE1;
-typedef unsigned long LONG1;
+typedef unsigned int LONG1;
 
 /* The SHS block size and message digest sizes, in bytes */
